finishing up

+++ src/com/donny/dendrofinance/account/Account.java
@@ -126,7 +126,7 @@ public class Account implements ExportableToJson, Serializable {
     public JsonObject export() throws JsonFormattingException {
         JsonObject obj = new JsonObject();
         obj.put("name", new JsonString(NAME));
-        obj.put("id", new JsonDecimal(BigDecimal.valueOf(AID)));
+        obj.put("id", new JsonDecimal(AID));
         obj.put("currency", new JsonString(CUR.toString()));
         obj.put("type", new JsonString(TYPE.NAME));
         if (EXCHANGE != null) {
