+++ src/com/donny/dendrofinance/gui/password/PasswordGui.java
@@ -207,7 +207,7 @@ public class PasswordGui extends JFrame {
         CURRENT_INSTANCE.day = false;
         config.put("flags", new JsonString("lxAd"));
         CURRENT_INSTANCE.precision = new MathContext(20);
-        config.put("precision", new JsonDecimal(BigDecimal.valueOf(20)));
+        config.put("precision", new JsonDecimal(20));
         CURRENT_INSTANCE.logLevel = new LogHandler.LogLevel("info");
         config.put("log", new JsonString("info"));
         CURRENT_INSTANCE.mainTicker = "USD";
